models/use-face-recognition/requirements.txt
face_recognition==1.3.0
flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.0
pillow==8.3.1
